web_services/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
filelock==3.0.12
googlemaps==4.2.2
idna==2.9
psycopg2==2.8.5
pytz==2019.3
requests==2.23.0
responses==0.10.14
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.18